# apps/messaging-service/src/routes/auth.ts
import { Router } from 'express';
import type { Env } from '../config/env.js';
import { issueAuthTokens } from '../auth/issueTokens.js';
import {
  signEmailVerificationToken,
  verifyEmailVerificationToken,
} from '../auth/emailVerificationJwt.js';
import {
  signPasswordResetToken,
  verifyPasswordResetToken,
} from '../auth/passwordResetJwt.js';
import { getRefreshPayload, revokeRefreshToken } from '../auth/refreshTokenRedis.js';
import { getClientIp } from '../auth/getClientIp.js';
import {
  emailRateLimitKey,
  rateLimitExceeded,
} from '../auth/rateLimitRedis.js';
import { AppError } from '../errors/AppError.js';
import { logger } from '../logger.js';
import { verifyPassword } from '../users/password.js';
import {
  createUser,
  DuplicateEmailError,
  findUserByEmail,
  findUserById,
  setUserEmailVerified,
  setUserPasswordAndBumpVersion,
} from '../users/repo.js';
import { toUserApiShape } from '../users/publicUser.js';
import type { RegisterRequest } from '../validation/schemas.js';
import { sendPasswordResetEmail } from '../email/sendPasswordResetEmail.js';
import { sendVerificationEmail } from '../email/sendVerificationEmail.js';
import { validateBody } from '../validation/middleware.js';
import {
  forgotPasswordRequestSchema,
  loginRequestSchema,
  logoutRequestSchema,
  refreshRequestSchema,
  registerRequestSchema,
  resendVerificationRequestSchema,
  resetPasswordRequestSchema,
  verifyEmailRequestSchema,
} from '../validation/schemas.js';

function requireJwtForAuth(env: Env): void {
  if (!env.JWT_SECRET?.trim()) {
    throw new AppError(
      'AUTH_NOT_CONFIGURED',
      503,
      'JWT_SECRET is not configured',
    );
  }
}

/** When verification is off, verify/resend are not part of the happy path — reject with a clear error. */
function requireEmailVerificationEnabled(env: Env): void {
  if (!env.EMAIL_VERIFICATION_REQUIRED) {
    throw new AppError(
      'EMAIL_VERIFICATION_DISABLED',
      400,
      'Email verification is not enabled (EMAIL_VERIFICATION_REQUIRED=false). Accounts are verified on registration.',
    );
  }
}

export function createAuthRouter(env: Env): Router {
  const router = Router();

  router.post(
    '/auth/register',
    validateBody(registerRequestSchema),
    async (req, res, next) => {
      try {
        requireJwtForAuth(env);
        const ip = getClientIp(req);
        const regKey = `ratelimit:register:ip:${ip}`;
        if (
          await rateLimitExceeded(
            regKey,
            env.REGISTER_RATE_LIMIT_WINDOW_SEC,
            env.REGISTER_RATE_LIMIT_MAX,
          )
        ) {
          next(
            new AppError(
              'RATE_LIMIT_EXCEEDED',
              429,
              'Too many registration attempts; try again later',
            ),
          );
          return;
        }

        const body = req.body as RegisterRequest;
        const emailVerified = !env.EMAIL_VERIFICATION_REQUIRED;
        const user = await createUser({
          email: body.email,
          password: body.password,
          profilePicture: body.profilePicture ?? undefined,
          status: body.status ?? undefined,
          emailVerified,
        });

        if (env.EMAIL_VERIFICATION_REQUIRED) {
          const verificationToken = await signEmailVerificationToken(
            env,
            user.id,
            user.email,
          );
          if (env.NODE_ENV !== 'production') {
            logger.info(
              { userId: user.id, email: user.email },
              'email verification token (dev — configure mail in production)',
            );
            logger.debug({ verificationToken }, 'verification JWT');
          } else {
            logger.info(
              { userId: user.id },
              'user registered; sending verification email when SendGrid is configured',
            );
          }
          try {
            await sendVerificationEmail(env, {
              to: user.email,
              verificationToken,
            });
          } catch (err: unknown) {
            logger.error(
              { err, userId: user.id },
              'failed to send verification email',
            );
          }
        } else {
          logger.info(
            { userId: user.id, email: user.email },
            'user registered; email verification skipped (EMAIL_VERIFICATION_REQUIRED=false)',
          );
        }

        if (!user.emailVerified) {
          res.status(201).json({
            accessToken: null,
            refreshToken: null,
            tokenType: 'Bearer',
            expiresIn: null,
          });
          return;
        }

        const tokens = await issueAuthTokens(env, user);
        res.status(201).json({
          accessToken: tokens.accessToken,
          refreshToken: tokens.refreshToken,
          tokenType: 'Bearer',
          expiresIn: tokens.expiresIn,
        });
      } catch (err: unknown) {
        if (err instanceof DuplicateEmailError) {
          next(
            new AppError(
              'EMAIL_ALREADY_REGISTERED',
              409,
              'An account with this email already exists',
            ),
          );
          return;
        }
        next(err);
      }
    },
  );

  router.post(
    '/auth/login',
    validateBody(loginRequestSchema),
    async (req, res, next) => {
      try {
        requireJwtForAuth(env);
        const body = req.body as { email: string; password: string };
        const user = await findUserByEmail(body.email);
        if (
          !user ||
          !(await verifyPassword(user.passwordHash, body.password))
        ) {
          next(
            new AppError(
              'INVALID_CREDENTIALS',
              401,
              'Invalid email or password',
            ),
          );
          return;
        }
        if (!user.emailVerified) {
          next(
            new AppError(
              'EMAIL_NOT_VERIFIED',
              403,
              'Verify your email before signing in',
            ),
          );
          return;
        }
        const tokens = await issueAuthTokens(env, user);
        res.status(200).json({
          accessToken: tokens.accessToken,
          refreshToken: tokens.refreshToken,
          tokenType: 'Bearer',
          expiresIn: tokens.expiresIn,
        });
      } catch (err: unknown) {
        next(err);
      }
    },
  );

  router.post(
    '/auth/refresh',
    validateBody(refreshRequestSchema),
    async (req, res, next) => {
      try {
        requireJwtForAuth(env);
        const { refreshToken: rawRefresh } = req.body as { refreshToken: string };
        const stored = await getRefreshPayload(rawRefresh);
        if (!stored) {
          next(
            new AppError(
              'INVALID_REFRESH_TOKEN',
              401,
              'Invalid or expired refresh token',
            ),
          );
          return;
        }
        const user = await findUserById(stored.userId);
        if (!user) {
          await revokeRefreshToken(rawRefresh);
          next(
            new AppError(
              'INVALID_REFRESH_TOKEN',
              401,
              'Invalid or expired refresh token',
            ),
          );
          return;
        }
        if ((user.refreshTokenVersion ?? 0) !== stored.v) {
          await revokeRefreshToken(rawRefresh);
          next(
            new AppError(
              'INVALID_REFRESH_TOKEN',
              401,
              'Refresh token has been revoked',
            ),
          );
          return;
        }
        if (env.EMAIL_VERIFICATION_REQUIRED && !user.emailVerified) {
          next(
            new AppError(
              'EMAIL_NOT_VERIFIED',
              403,
              'Verify your email before refreshing',
            ),
          );
          return;
        }
        const tokens = await issueAuthTokens(env, user);
        await revokeRefreshToken(rawRefresh);
        res.status(200).json({
          accessToken: tokens.accessToken,
          refreshToken: tokens.refreshToken,
          tokenType: 'Bearer',
          expiresIn: tokens.expiresIn,
        });
      } catch (err: unknown) {
        next(err);
      }
    },
  );

  router.post(
    '/auth/logout',
    validateBody(logoutRequestSchema),
    async (req, res, next) => {
      try {
        requireJwtForAuth(env);
        const { refreshToken: rawRefresh } = req.body as { refreshToken: string };
        await revokeRefreshToken(rawRefresh);
        res.status(204).send();
      } catch (err: unknown) {
        next(err);
      }
    },
  );

  router.post(
    '/auth/forgot-password',
    validateBody(forgotPasswordRequestSchema),
    async (req, res, next) => {
      try {
        requireJwtForAuth(env);
        const ip = getClientIp(req);
        const key = `ratelimit:forgot-password:ip:${ip}`;
        if (
          await rateLimitExceeded(
            key,
            env.FORGOT_PASSWORD_RATE_LIMIT_WINDOW_SEC,
            env.FORGOT_PASSWORD_RATE_LIMIT_MAX,
          )
        ) {
          next(
            new AppError(
              'RATE_LIMIT_EXCEEDED',
              429,
              'Too many password reset requests; try again later',
            ),
          );
          return;
        }
        const { email } = req.body as { email: string };
        const user = await findUserByEmail(email);
        if (user) {
          const token = await signPasswordResetToken(env, user.id, user.email);
          if (env.NODE_ENV !== 'production') {
            logger.debug(
              { userId: user.id },
              'password reset token (dev — do not log in production)',
            );
            logger.debug({ token }, 'password reset JWT');
          }
          try {
            await sendPasswordResetEmail(env, { to: user.email, token });
          } catch (err: unknown) {
            logger.error(
              { err, userId: user.id },
              'failed to send password reset email',
            );
          }
        }
        res.status(200).json({ ok: true });
      } catch (err: unknown) {
        next(err);
      }
    },
  );

  router.post(
    '/auth/reset-password',
    validateBody(resetPasswordRequestSchema),
    async (req, res, next) => {
      try {
        requireJwtForAuth(env);
        const body = req.body as { token: string; password: string };
        const claims = await verifyPasswordResetToken(env, body.token);
        const user = await findUserById(claims.userId);
        if (!user || user.email !== claims.email) {
          next(
            new AppError(
              'INVALID_PASSWORD_RESET_TOKEN',
              400,
              'Password reset token is not valid for this account',
            ),
          );
          return;
        }
        const ok = await setUserPasswordAndBumpVersion(user.id, body.password);
        if (!ok) {
          next(new AppError('INTERNAL_ERROR', 500, 'Could not update password'));
          return;
        }
        res.status(204).send();
      } catch (err: unknown) {
        next(err);
      }
    },
  );

  router.post(
    '/auth/verify-email',
    validateBody(verifyEmailRequestSchema),
    async (req, res, next) => {
      try {
        requireJwtForAuth(env);
        requireEmailVerificationEnabled(env);
        const ip = getClientIp(req);
        const key = `ratelimit:verify-email:ip:${ip}`;
        if (
          await rateLimitExceeded(
            key,
            env.VERIFY_EMAIL_RATE_LIMIT_WINDOW_SEC,
            env.VERIFY_EMAIL_RATE_LIMIT_MAX,
          )
        ) {
          next(
            new AppError(
              'RATE_LIMIT_EXCEEDED',
              429,
              'Too many attempts; try again later',
            ),
          );
          return;
        }

        const { token } = req.body as { token: string };
        const claims = await verifyEmailVerificationToken(env, token);

        const user = await findUserById(claims.userId);
        if (!user) {
          next(
            new AppError(
              'INVALID_VERIFICATION_TOKEN',
              400,
              'User not found for this token',
            ),
          );
          return;
        }
        if (user.email !== claims.email) {
          next(
            new AppError(
              'INVALID_VERIFICATION_TOKEN',
              400,
              'Token does not match user email',
            ),
          );
          return;
        }
        if (user.emailVerified) {
          const tokens = await issueAuthTokens(env, user);
          res.status(200).json({
            user: toUserApiShape(user),
            accessToken: tokens.accessToken,
            refreshToken: tokens.refreshToken,
            tokenType: 'Bearer',
            expiresIn: tokens.expiresIn,
          });
          return;
        }

        await setUserEmailVerified(user.id, true);
        const updated = await findUserById(user.id);
        if (!updated) {
          next(
            new AppError('INTERNAL_ERROR', 500, 'User missing after verify'),
          );
          return;
        }
        const tokens = await issueAuthTokens(env, updated);
        res.status(200).json({
          user: toUserApiShape(updated),
          accessToken: tokens.accessToken,
          refreshToken: tokens.refreshToken,
          tokenType: 'Bearer',
          expiresIn: tokens.expiresIn,
        });
      } catch (err: unknown) {
        next(err);
      }
    },
  );

  router.post(
    '/auth/resend-verification',
    validateBody(resendVerificationRequestSchema),
    async (req, res, next) => {
      try {
        requireJwtForAuth(env);
        requireEmailVerificationEnabled(env);
        const body = req.body as { email: string };
        const emailKey = emailRateLimitKey(body.email);
        if (
          await rateLimitExceeded(
            emailKey,
            env.RESEND_RATE_LIMIT_WINDOW_SEC,
            env.RESEND_RATE_LIMIT_MAX,
          )
        ) {
          next(
            new AppError(
              'RATE_LIMIT_EXCEEDED',
              429,
              'Too many resend requests for this email; try again later',
            ),
          );
          return;
        }

        const user = await findUserByEmail(body.email);
        if (!user || user.emailVerified) {
          res.status(200).json({ ok: true });
          return;
        }

        const verificationToken = await signEmailVerificationToken(
          env,
          user.id,
          user.email,
        );
        if (env.NODE_ENV !== 'production') {
          logger.debug(
            { userId: user.id, verificationToken },
            'resend verification JWT (dev)',
          );
        } else {
          logger.info(
            { userId: user.id },
            'resend verification email when SendGrid is configured',
          );
        }

        try {
          await sendVerificationEmail(env, {
            to: user.email,
            verificationToken,
          });
        } catch (err: unknown) {
          logger.error(
            { err, userId: user.id },
            'failed to send verification resend email',
          );
          next(
            new AppError(
              'EMAIL_SEND_FAILED',
              503,
              'Could not send verification email; try again later',
            ),
          );
          return;
        }

        res.status(200).json({ ok: true });
      } catch (err: unknown) {
        next(err);
      }
    },
  );

  return router;
}