# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Messaging API
  version: 0.1.6
  description: |
    **Source of truth** for REST contracts between **web-client** and **messaging-service**
    (single HTTP API; there is no separate notification REST service).

    **Real-time behaviour** — chat, call signaling, and **in-tab notification events** — uses
    **Socket.IO** on **messaging-service**, not additional REST resources here. The server emits
    **`notification`** with a JSON body discriminated by **`kind`** (`message` \| `call_incoming`);
    see **`docs/PROJECT_PLAN.md`** §8 (and §3.3). The web-client runs
    **`socket.io-client` in a Web Worker** and forwards events to the UI; that is a client
    concern and is not described by this OpenAPI document.

    All paths below are served under the **`/v1`** prefix (see `servers`). Implementations in
    **messaging-service** must stay aligned; regenerate **`apps/web-client`** types with
    `npm run generate:api`. Routes marked *planned* in summaries are not necessarily wired yet.

    **Users:** `User` includes optional **profile picture** and **status** (short text).
    At **signup** (`POST /auth/register`), only **`email`** and **`password`** are required; **`profilePicture`**
    (URL) and **`status`** are **optional** — avatar URL may come from a prior **`POST /media/upload`**
    if the client uploads before register. **Update profile** — **`PATCH /users/me`** accepts
...
    **`displayName`** parts.
...

...

...
servers:
  - url: /v1
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      summary: Register a new user (Argon2 password hash; optional profilePicture and status)
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: >-
            Created. **`accessToken`** is **`null`** when **`emailVerified`** is false (email verification required);
            use **`POST /auth/verify-email`** then **`accessToken`** is returned there, or sign in with **`POST /auth/login`** after verification.
            When verification is not required, an access token is returned immediately.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many registration attempts from this client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Auth signing not configured (e.g. missing JWT_SECRET)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/verify-email:
    post:
      operationId: verifyEmail
      summary: Verify email using signed JWT from registration or resend (throttled per IP)
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified (or already verified) — includes **`accessToken`** for the verified session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyEmailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          description: Too many verification attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Auth signing not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/resend-verification:
    post:
      operationId: resendVerificationEmail
      summary: Resend verification flow (throttled per email; constant response to avoid enumeration)
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
      responses:
        '200':
          description: >-
            Request accepted (always ok true in body whether or not an unverified user existed).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResendVerificationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          description: Too many resend requests for this email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Auth signing not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      operationId: login
      summary: Email/password login
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Access and refresh tokens issued (**only** when **`emailVerified`** is true)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Email not verified — complete **`POST /auth/verify-email`** first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      operationId: refreshTokens
      summary: Rotate refresh token and issue new access token
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New access and refresh tokens (previous refresh token invalidated)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Email not verified — cannot refresh
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Auth signing not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      operationId: logout
      summary: Revoke refresh token (logout)
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: Refresh token revoked
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          description: Auth signing not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      summary: Request password reset email (same response whether or not email exists)
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          description: Too many requests from this client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Auth signing not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/reset-password:
    post:
      operationId: resetPassword
      summary: Set new password using JWT from forgot-password email
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Password updated; all refresh tokens invalidated server-side
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          description: Auth signing not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token (see Auth responses for refresh flow)
...
  responses:
    BadRequest:
      description: Validation or malformed request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Stable machine-readable error code
        message:
          type: string
          description: Safe client-facing message
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        profilePicture:
          type: string
          format: uri
          nullable: true
          description: Optional avatar URL at signup (e.g. from `POST /media/upload` before register)
        status:
          type: string
          nullable: true
          maxLength: 280
          description: Optional short status line at signup
    VerifyEmailRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: Signed JWT from registration or resend
    VerifyEmailResponse:
      type: object
      required:
        - user
        - accessToken
        - refreshToken
        - tokenType
        - expiresIn
      properties:
        user:
          $ref: '#/components/schemas/User'
        accessToken:
          type: string
          description: Access JWT for the verified session
        refreshToken:
          type: string
          description: Opaque refresh token (stored in Redis)
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          description: Access token lifetime in seconds
    ResendVerificationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    ResendVerificationResponse:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    LogoutRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required:
        - token
        - password
      properties:
        token:
          type: string
          description: Signed JWT from forgot-password email
        password:
          type: string
          format: password
          minLength: 8
    OkResponse:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
    AuthResponse:
      type: object
      required:
        - tokenType
      properties:
        accessToken:
          type: string
          nullable: true
          description: null after register when email verification is pending; otherwise HS256 JWT
        refreshToken:
          type: string
          nullable: true
          description: Opaque refresh token (Redis); null when access token is not issued
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          nullable: true
          description: null when accessToken is null; otherwise access token lifetime in seconds
    User:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
          nullable: true
        emailVerified:
          type: boolean
        profilePicture:
          type: string
          format: uri
          nullable: true
          description: HTTPS URL or CDN URL for avatar (or null if unset)
        status:
          type: string
          nullable: true
          maxLength: 280
          description: Short user-defined status line
...